video-backend/requirements.txt
flask==3.1.0
boto3==1.35.0
python-dotenv==1.0.1
gunicorn==23.0.0
PyPDF2==3.0.0
pdf2image==1.16.0